pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    # Existing dependencies
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-generativeai>=0.8.4",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "replit>=4.1.1",
    "sendgrid>=6.11.0",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "flask-login>=0.6.3",
    
    # Phase 1: Supabase & AI Infrastructure
    "supabase>=2.4.0",
    "python-dotenv>=1.0.0",
    "stripe>=7.0.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
]
